+++ src/pages/app/orders/orders.tsx
@@ -1,8 +1,7 @@
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
-import { ArrowRight, Search, X } from 'lucide-react';
+import { Table, TableBody, TableHead, TableHeader, TableRow } from '@/components/ui/table';
 import { Helmet } from 'react-helmet-async';
+import { OrderTabelRow } from './order-tabel-row';
+import { OrderTableFilters } from './order-table-filters';
 
 export const Orders = () => {
   return (
@@ -12,10 +11,7 @@ export const Orders = () => {
         <h1 className="text-3xl font-bold tracking-tight">Pedidos</h1>
       </div>
       <div className="space-y-2 5">
-        <form className="flex items-center gap-2">
-          <span className="text-sm font-semibold">Filtros</span>
-          <Input placeholder="Nome do cliente" className="h-8 w-[320px]" />
-        </form>
+        <OrderTableFilters />
 
         <div className="border rounded-md">
           <Table>
@@ -33,37 +29,7 @@ export const Orders = () => {
             </TableHeader>
             <TableBody>
               {Array.from({ length: 10 }).map((_, i) => (
-                <TableRow key={i}>
-                  <TableCell>
-                    <Button variant="outline" size="sm">
-                      <Search className="h-3 w-3" />
-                      <span className="sr-only">Detalhes do pedido</span>
-                    </Button>
-                  </TableCell>
-                  <TableCell className="font-mono text-xs font-medium">4154asd5</TableCell>
-                  <TableCell className="text-muted-foreground">há 10 minutos</TableCell>
-                  <TableCell>
-                    <div className="flex items-center gap-2">
-                      <span className="h-2 w-2 rounded-full bg-slate-400" />
-                      <span className="font-medium text-muted-foreground">Pendente</span>
-                    </div>
-                  </TableCell>
-                  <TableCell className="font-medium">João da Silva</TableCell>
-                  <TableCell className="font-medium">R$ 150,00</TableCell>
-
-                  <TableCell>
-                    <Button variant="ghost" size="sm">
-                      <ArrowRight className="h-3 w-3 mr-2" />
-                      Aprovar
-                    </Button>
-                  </TableCell>
-                  <TableCell>
-                    <Button variant="ghost" size="sm">
-                      <X className="h-3 w-3 mr-2" />
-                      Cancelar
-                    </Button>
-                  </TableCell>
-                </TableRow>
+                <OrderTabelRow key={i} />
               ))}
             </TableBody>
           </Table>